MILP model formulations & Data
All files are arranged in folders. The data folder contains data for the 2014-2018 Championship Tour Seasons. The 2017MILP are all MILP team-selection strategies for the 2017 season. The 2018MILP are all MILP team-selection strategies for the 2018 season. The 2017 and 2018 leaderboard contains that ranking of all surfers pre- and post-event on tye 2017 and 2018 Championship tour seasons.